refactor(app/cmd): コマンド別にファイルを分離
